feat(auth,users): JWT com refresh + blacklist, RBAC e novo modelo de usuários" -m "$(cat << 'MSG' - Implementa autenticação JWT com **access/refresh tokens**, rotação de refresh e **blacklist** no logout - Refatora camada de usuários para modelo Party → Person → User; perfis Teacher/Student e RBAC - Controllers com **@PreAuthorize (owner-or-admin)** e SecurityConfig stateless (JWT) - Seeds, Exception handling consistente e coleção **Postman** + **Environments**
### Auth / Security
- **config/SecurityConfig**: API stateless, CORS, `@EnableMethodSecurity`, regras públicas e protegidas;
  inclui `DaoAuthenticationProvider` (BCrypt) + encadeado no `HttpSecurity`.
- **filters/JwtAuthenticationFilter**: alinha rotas públicas, checa **blacklist**, rejeita refresh em rotas não-/auth.
- **filters/CustomUserPrincipal**: principal com `id` exposto para SpEL (`principal.id`).
- **filters/CustomUserDetailsService**: carrega usuário por e-mail, mapeia roles para `ROLE_*`.
- **services/JwtService**: separa `generateAccessToken` / `generateRefreshToken`, adiciona claim `typ`, `isRefreshToken`,
  `extractExpiration`, compat com métodos antigos.
- **services/TokenBlacklistService**: blacklist in-memory com expiração (pronto para trocar por Redis).

### Controllers
- **controllers/AuthController**:
  - `POST /auth/login` → emite access+refresh
  - `POST /auth/refresh` → valida/rota refresh, blacklist do antigo
  - `POST /auth/logout` → inclui access/refresh na blacklist
- **controllers/UserController**: CRUD + updates (email, senha, status, person) + gestão de roles,
  com `@PreAuthorize` (dono ou admin).
- **controllers/ProfileController**: CRUD de **TeacherProfile** e **StudentProfile** com `@PreAuthorize`
  (público apenas `GET /profiles/teacher/{userId}`).

### Services / Domain
- **services/UserService**: create/list/get/update/delete; roles add/remove/replace; person updates.
- **services/TeacherProfileService**: CRUD + helpers de especialidades (add/remove/replace).
- **services/StudentProfileService**: CRUD.
- **services/RoleService**: utilizado para garantir roles (já integrado nos services de perfil).

### DTOs
- **dtos/UserDtos**: inclui `UserDetailsResponse`, requests de update (`UpdateEmailRequest`, `UpdatePasswordRequest`,
  `UpdateStatusRequest`, `UpdatePersonDataRequest`), `RolesRequest` e DTOs de perfis (create/update/response).

### Infra / Boot
- **config/GlobalExceptionHandler (@RestControllerAdvice)**: respostas padronizadas (400/401/403/404/409/500)
  com `ProblemDetail` para `ErrorResponseException`.
- **config/DataInitializer**: garante roles; opcionalmente cria admin e usuários demo (teacher/student),
  com propriedades `app.seed.*`.

### Postman
- **postman/Quadrago.postman_collection.json**: Auth (login/refresh/logout), Users e Profiles (Teacher/Student),
  com testes que preenchem `accessToken`, `refreshToken`, `userId`.
- **postman/Quadrago_-_Local_Dev.postman_environment.json**
- **postman/Quadrago_-_Staging.postman_environment.json**
- **postman/Quadrago_-_Production.postman_environment.json**

### Propriedades (application.properties) — adicionar/ajustar
- `spring.jpa.open-in-view=false`
- `spring.jpa.properties.hibernate.default_batch_fetch_size=50`
- `spring.jpa.properties.hibernate.jdbc.time_zone=UTC`
- `spring.jackson.time-zone=UTC`
- **JWT**:
  - `jwt.secret=...` (>= 32 bytes ou Base64)
  - `jwt.access.expiration=900000`   # 15m
  - `jwt.refresh.expiration=2592000000` # 30d
- **Seeds**:
  - `app.seed.admin.enabled=true`
  - `app.seed.admin.email=admin@quadrago.local`
  - `app.seed.admin.password=admin`
  - `app.seed.demo.enabled=false`

### Notas
- Verifique se os pacotes batem com: `com.quadrago.backend.{config,controllers,filters,services,dtos,...}`.
- Em produção, troque CORS `*` por domínios do front.
- Se desejar Redis para blacklist, basta implementar um `TokenBlacklistService` com TTL.

BREAKING CHANGE: Regras de autorização e rotas atualizadas para `/users` e `/profiles`.
MSG
)

+++ src/main/java/com/quadrago/backend/repositories/RoleRepository.java
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
-
+public interface RoleRepository extends JpaRepository<Role, Short> {
     Optional<Role> findByName(RoleName name);
 }
